feat(backend): ADR-0008 governance Decimal migration — units/meetings/age_request/etat_date/vote/resolution (closes #525, #534 Story C1)
Belgian copro quorum/voting-power/quota arithmetic now exact
rust_decimal::Decimal end-to-end (ADR-0008). Closes the #525
ColumnDecode panic (units.quota DOUBLE PRECISION read into
Unit.quota: Decimal).

SQL migration 20260516000000_alter_governance_to_numeric.sql:
- units.quota DOUBLE PRECISION -> NUMERIC(10,4)
- meetings.total_quotas / present_quotas DOUBLE PRECISION ->
  NUMERIC(10,4) (nullable preserved)

Entities f64 -> Decimal:
- Vote.voting_power; Resolution.total_voting_power_{pour,contre,
  abstention} + record_*/calculate_result/close_voting
- Meeting.{total_quotas,present_quotas}; validate_quorum now uses
  EXACT decimal compare present > total/2 (Art. 3.87 §5 CC) — no
  IEEE754 rounding at the legal boundary (quorum_percentage stays
  f64, display-only derived value)
- AgeRequestCosignatory.shares_pct; AgeRequest.{total_shares_pct,
  threshold_pct,DEFAULT_THRESHOLD_PCT}; shares_pct_missing -> Decimal

Repos: removed ::FLOAT8 casts / get::<f64> in vote_repository_impl,
resolution_repository_impl, age_request_repository_impl (columns
already DECIMAL/NUMERIC -> map straight to Decimal, no f64 round
trip). unit_repository_impl fixed purely by the SQL migration.

Ports + use_cases + DTOs cascaded (Decimal serializes as JSON
string = exact, consistent with Story A). meeting_minutes_exporter
sums voting power in Decimal.

Validation: cargo check --all-targets clean (0 errors). 1217 lib
unit tests pass, 0 failed — incl. all migrated entity/use-case
tests and exact-quorum-boundary tests
(test_quorum_not_reached_at_50_percent_exact). The new
governance_decimal.feature BDD scenarios cannot run locally:
bdd_governance setup_database() hardcodes 127.0.0.1 for the
testcontainers Postgres, so ALL 22 governance scenarios fail
identically with PoolTimedOut inside the dev container
(pre-existing infra limit, out of Story C1 scope) — to be
validated on CI host runner.

Scope notes:
- age_request stack kept Result<_, String> (entity-pervasive;
  migrating to AppError would balloon scope beyond C1) — noted
  per brief.
- validate_proxy_mandate left i32/f64 (transient computed ratio,
  not persisted NUMERIC; not legally load-bearing).
- IoT/energy untouched (ADR-0009). payment_reminder/budget/stats
  untouched (Story C2).

Closes #525
Refs #534, #521

+++ backend/src/application/ports/vote_repository.rs
@@ -1,5 +1,6 @@
 use crate::domain::entities::Vote;
 use async_trait::async_trait;
+use rust_decimal::Decimal;
 use uuid::Uuid;
 
 /// Port (trait) for Vote repository operations
@@ -43,7 +44,7 @@ pub trait VoteRepository: Send + Sync {
     async fn sum_voting_power_by_resolution(
         &self,
         resolution_id: Uuid,
-    ) -> Result<(f64, f64, f64), String>; // (pour, contre, abstention)
+    ) -> Result<(Decimal, Decimal, Decimal), String>; // (pour, contre, abstention)
 
     /// Count proxy votes held by a mandataire on a given resolution (Art. 3.87 §7 CC).
     /// Returns (count_proxies, total_proxy_voting_power).
@@ -51,5 +52,5 @@ pub trait VoteRepository: Send + Sync {
         &self,
         resolution_id: Uuid,
         proxy_owner_id: Uuid,
-    ) -> Result<(i64, f64), String>;
+    ) -> Result<(i64, Decimal), String>;
 }
